Move the page processing code to its own module and covnert to TypeScript
